refactor: extract shared TUI framework — tui_common.rs with RepoRow, KeyAction, OverlayState + unified rendering
- Create tui_common.rs: shared types (RepoRow, KeyAction, OverlayState) and
  rendering functions (header, table, detail, footer, overlay, backdrop,
  centered_modal, handle_key, is_quit_key, pulse_bright) used by both
  embedded and remote TUI
- Embedded tui.rs: reduced from ~1380 to ~580 lines, imports from tui_common,
  keeps only serve-specific logic (map_repo_rows, build_info_overlay,
  spawn_doctor, spawn_force_reindex, helpers)
- Remote tui_remote.rs: reduced from ~610 to ~340 lines, imports from
  tui_common, actions via HTTP endpoints (GET /repos/{alias}/info,
  POST /repos/{alias}/doctor)
- serve/mod.rs: add /repos/{alias}/info and /repos/{alias}/doctor HTTP
  endpoints for remote TUI actions
- Net -34 lines while adding new functionality (i/d/f keys for remote TUI)
