+++ src/serve/tui_remote.rs
@@ -1,27 +1,28 @@
 //! Standalone TUI that connects to a running `codesearch serve` via HTTP.
 //!
 //! Polls `GET /status` every second and renders the same ratatui layout
-//! as the embedded TUI in `tui.rs`. This allows the TUI to be opened and
-//! closed independently of the server process.
+//! as the embedded TUI in `tui.rs`, using the shared `tui_common` framework.
+//! This allows the TUI to be opened and closed independently of the server process.
+//!
+//! Actions (i/d/f/s) are routed to HTTP endpoints on the serve.
 
 use std::io;
 use std::time::Duration;
 
 use anyhow::Result;
 use ratatui::backend::CrosstermBackend;
-use ratatui::layout::{Constraint, Layout, Margin, Rect};
-use ratatui::style::{Color, Modifier, Style};
-use ratatui::text::{Line, Span};
-use ratatui::widgets::{Block, Borders, Cell, Row, Table, TableState};
+use ratatui::layout::{Constraint, Layout};
 use ratatui::Terminal;
 
-use crossterm::event::{self, Event, KeyCode, KeyEvent, KeyEventKind};
+use crossterm::event::{self, Event, KeyEventKind};
 use crossterm::terminal::{self, EnterAlternateScreen, LeaveAlternateScreen};
 
 use serde::Deserialize;
 
+use super::tui_common::{self, KeyAction, OverlayState, RepoRow};
+
 // ---------------------------------------------------------------------------
-// Data model for the remote TUI
+// Data model (HTTP JSON response)
 // ---------------------------------------------------------------------------
 
 #[derive(Debug, Default, Deserialize)]
@@ -42,12 +43,26 @@ struct RepoInfo {
     last_tool_call: Option<String>,
     tool_call_count: u64,
     csharp_index: String,
+    #[serde(default)]
+    csharp_error: Option<String>,
+    #[serde(default)]
+    path: String,
 }
 
-#[derive(Debug, Default)]
-struct RemoteTuiState {
-    data: Option<StatusResponse>,
-    connection_errors: u32,
+impl RepoInfo {
+    fn to_row(&self) -> RepoRow {
+        RepoRow {
+            alias: self.alias.clone(),
+            status: self.status.clone(),
+            csharp_index: self.csharp_index.clone(),
+            csharp_error: self.csharp_error.clone(),
+            changes: self.changes,
+            tool_call_count: self.tool_call_count,
+            last_tool_call: self.last_tool_call.clone(),
+            lock_mode: self.lock_mode.clone(),
+            path: self.path.clone(),
+        }
+    }
 }
 
 // ---------------------------------------------------------------------------
@@ -80,43 +95,64 @@ async fn run_remote_tui_loop(
     terminal: &mut Terminal<CrosstermBackend<io::Stdout>>,
     serve_url: &str,
 ) -> Result<()> {
-    let mut table_state = TableState::default();
+    let mut table_state = ratatui::widgets::TableState::default();
     table_state.select(Some(0));
 
-    let mut remote_state = RemoteTuiState::default();
+    let mut data: Option<StatusResponse> = None;
+    let mut connection_errors: u32 = 0;
 
     let status_url = format!("{}/status", serve_url.trim_end_matches('/'));
     let poll_interval = Duration::from_secs(1);
     let key_poll_timeout = Duration::from_millis(100);
 
+    // Optional overlay
+    let mut overlay: Option<OverlayState> = None;
+
+    // Channel for async doctor results
+    let (doctor_tx, mut doctor_rx) = tokio::sync::mpsc::channel::<OverlayState>(1);
+
     loop {
         // Fetch status from serve
         match reqwest::get(&status_url).await {
             Ok(resp) if resp.status().is_success() => match resp.json::<StatusResponse>().await {
-                Ok(data) => {
-                    remote_state.data = Some(data);
-                    remote_state.connection_errors = 0;
+                Ok(d) => {
+                    data = Some(d);
+                    connection_errors = 0;
                 }
                 Err(_) => {
-                    remote_state.connection_errors += 1;
+                    connection_errors += 1;
                 }
             },
             _ => {
-                remote_state.connection_errors += 1;
+                connection_errors += 1;
             }
         }
 
+        // Build RepoRows from HTTP data
+        let rows: Vec<RepoRow> = data
+            .as_ref()
+            .map(|d| d.repos.iter().map(|r| r.to_row()).collect())
+            .unwrap_or_default();
+
         // Clamp selection
-        if let Some(ref data) = remote_state.data {
-            if !data.repos.is_empty() {
-                let sel = table_state.selected().unwrap_or(0);
-                if sel >= data.repos.len() {
-                    table_state.select(Some(data.repos.len() - 1));
-                }
+        if !rows.is_empty() {
+            let sel = table_state.selected().unwrap_or(0);
+            if sel >= rows.len() {
+                table_state.select(Some(rows.len() - 1));
             }
+        }
+
+        // Check if doctor result arrived
+        if let Ok(result) = doctor_rx.try_recv() {
+            overlay = Some(result);
         }
 
         // Render
+        let version = data.as_ref().map(|d| d.version.as_str()).unwrap_or("?");
+        let active = data.as_ref().map(|d| d.active_sessions).unwrap_or(0);
+        let cpu = data.as_ref().map(|d| d.cpu_percent.as_str()).unwrap_or("—");
+        let csharp_helper = data.as_ref().map(|d| d.csharp_helper).unwrap_or(false);
+
         terminal.draw(|f| {
             let size = f.area();
             let chunks = Layout::vertical([
@@ -127,36 +163,39 @@ async fn run_remote_tui_loop(
             ])
             .split(size);
 
-            match &remote_state.data {
-                Some(data) => {
-                    render_header(f, chunks[0], serve_url, &data.version);
-                    render_table(f, chunks[1], &data.repos, &mut table_state);
-                    render_detail(f, chunks[2], &data.repos, &table_state);
-                    render_footer(
-                        f,
-                        chunks[3],
-                        &data.repos,
-                        &table_state,
-                        data.active_sessions,
-                        &data.cpu_percent,
-                        data.csharp_helper,
-                    );
-                }
-                None => {
-                    render_header(f, chunks[0], serve_url, "?");
-                    let msg = if remote_state.connection_errors >= 3 {
-                        "Connection lost — will retry..."
-                    } else {
-                        "Connecting..."
-                    };
-                    let connecting =
-                        ratatui::widgets::Paragraph::new(Line::from(vec![Span::styled(
-                            format!(" {} ", msg),
-                            Style::default().fg(Color::Yellow),
-                        )]));
-                    f.render_widget(connecting, chunks[1]);
-                    render_footer(f, chunks[3], &[], &table_state, 0, "—", false);
-                }
+            if data.is_some() {
+                tui_common::render_header(f, chunks[0], serve_url, version, true);
+                tui_common::render_table(f, chunks[1], &rows, &mut table_state);
+                tui_common::render_detail(f, chunks[2], &rows, &table_state, 3);
+                tui_common::render_footer(
+                    f,
+                    chunks[3],
+                    &rows,
+                    &table_state,
+                    active,
+                    cpu,
+                    csharp_helper,
+                );
+            } else {
+                tui_common::render_header(f, chunks[0], serve_url, "?", true);
+                let msg = if connection_errors >= 3 {
+                    "Connection lost — will retry..."
+                } else {
+                    "Connecting..."
+                };
+                let connecting = ratatui::widgets::Paragraph::new(ratatui::text::Line::from(vec![
+                    ratatui::text::Span::styled(
+                        format!(" {} ", msg),
+                        ratatui::style::Style::default().fg(ratatui::style::Color::Yellow),
+                    ),
+                ]));
+                f.render_widget(connecting, chunks[1]);
+                tui_common::render_footer(f, chunks[3], &[], &table_state, 0, "—", false);
+            }
+
+            // Render overlay on top if active
+            if let Some(ref ov) = overlay {
+                tui_common::render_overlay(f, size, ov);
             }
         })?;
 
@@ -167,24 +206,163 @@ async fn run_remote_tui_loop(
                 if key.kind != KeyEventKind::Press {
                     continue;
                 }
-                if is_quit_key(key) {
+
+                // If overlay is active, Esc dismisses it; no other keys processed
+                if overlay.is_some() {
+                    if matches!(key.code, crossterm::event::KeyCode::Esc) {
+                        overlay = None;
+                    }
+                    continue;
+                }
+
+                if tui_common::is_quit_key(key) {
                     should_quit = true;
                     break;
                 }
-                let repo_count = remote_state
-                    .data
-                    .as_ref()
-                    .map(|d| d.repos.len())
-                    .unwrap_or(0);
-                if handle_key(key, &mut table_state, repo_count) {
-                    // 's' pressed — trigger reload via HTTP
-                    let reload_url = format!("{}/reload", serve_url.trim_end_matches('/'));
-                    // Fire-and-forget POST; don't block the TUI loop
-                    let _ = reqwest::Client::new()
-                        .post(&reload_url)
-                        .timeout(std::time::Duration::from_secs(3))
-                        .send()
-                        .await;
+
+                let action = tui_common::handle_key(key, &mut table_state, rows.len());
+                match action {
+                    KeyAction::Reload => {
+                        let reload_url = format!("{}/reload", serve_url.trim_end_matches('/'));
+                        let _ = reqwest::Client::new()
+                            .post(&reload_url)
+                            .timeout(Duration::from_secs(3))
+                            .send()
+                            .await;
+                    }
+                    KeyAction::ShowInfo(idx) => {
+                        if idx < rows.len() {
+                            let alias = rows[idx].alias.clone();
+                            // Fetch info via HTTP
+                            let info_url =
+                                format!("{}/repos/{}/info", serve_url.trim_end_matches('/'), alias);
+                            overlay = Some(OverlayState::DoctorRunning {
+                                alias: alias.clone(),
+                            });
+                            let tx = doctor_tx.clone();
+                            tokio::spawn(async move {
+                                let result = match reqwest::Client::new()
+                                    .get(&info_url)
+                                    .timeout(Duration::from_secs(10))
+                                    .send()
+                                    .await
+                                {
+                                    Ok(resp) if resp.status().is_success() => {
+                                        match resp.json::<InfoResponse>().await {
+                                            Ok(info) => OverlayState::Info {
+                                                alias,
+                                                chunks: info.chunks,
+                                                files: info.files,
+                                                max_chunk_id: info.max_chunk_id,
+                                                db_size_human: info.db_size_human,
+                                                model: info.model,
+                                                dims: info.dims,
+                                                lock: info.lock,
+                                                index_age: info.index_age,
+                                            },
+                                            Err(e) => OverlayState::Doctor {
+                                                alias,
+                                                results: vec![
+                                                    format!("✗ Failed to parse info: {}", e),
+                                                    String::new(),
+                                                    "  [Esc] close".to_string(),
+                                                ],
+                                            },
+                                        }
+                                    }
+                                    _ => {
+                                        // Fallback: show basic info
+                                        let _ = tx; // suppress unused warning
+                                        OverlayState::Doctor {
+                                            alias,
+                                            results: vec![
+                                                "✗ Info endpoint not available".to_string(),
+                                                "  (Upgrade serve to get /repos/{alias}/info)"
+                                                    .to_string(),
+                                                String::new(),
+                                                "  [Esc] close".to_string(),
+                                            ],
+                                        }
+                                    }
+                                };
+                                let _ = tx.send(result).await;
+                            });
+                        }
+                    }
+                    KeyAction::RunDoctor(idx) => {
+                        if idx < rows.len() {
+                            let alias = rows[idx].alias.clone();
+                            overlay = Some(OverlayState::DoctorRunning {
+                                alias: alias.clone(),
+                            });
+                            let tx = doctor_tx.clone();
+                            let serve = serve_url.to_string();
+                            tokio::spawn(async move {
+                                let doctor_url = format!(
+                                    "{}/repos/{}/doctor",
+                                    serve.trim_end_matches('/'),
+                                    alias
+                                );
+                                let result = match reqwest::Client::new()
+                                    .post(&doctor_url)
+                                    .timeout(Duration::from_secs(60))
+                                    .send()
+                                    .await
+                                {
+                                    Ok(resp) if resp.status().is_success() => {
+                                        match resp.json::<DoctorResponse>().await {
+                                            Ok(d) => OverlayState::Doctor {
+                                                alias,
+                                                results: d.results,
+                                            },
+                                            Err(e) => OverlayState::Doctor {
+                                                alias,
+                                                results: vec![
+                                                    format!("✗ Failed to parse doctor: {}", e),
+                                                    String::new(),
+                                                    "  [Esc] close".to_string(),
+                                                ],
+                                            },
+                                        }
+                                    }
+                                    Ok(resp) => OverlayState::Doctor {
+                                        alias,
+                                        results: vec![
+                                            format!("✗ Doctor failed: HTTP {}", resp.status()),
+                                            String::new(),
+                                            "  [Esc] close".to_string(),
+                                        ],
+                                    },
+                                    Err(e) => OverlayState::Doctor {
+                                        alias,
+                                        results: vec![
+                                            format!("✗ Doctor request failed: {}", e),
+                                            String::new(),
+                                            "  [Esc] close".to_string(),
+                                        ],
+                                    },
+                                };
+                                let _ = tx.send(result).await;
+                            });
+                        }
+                    }
+                    KeyAction::ForceReindex(idx) => {
+                        if idx < rows.len() {
+                            let alias = rows[idx].alias.clone();
+                            let reindex_url = format!(
+                                "{}/repos/{}/reindex?force=true",
+                                serve_url.trim_end_matches('/'),
+                                alias
+                            );
+                            // Fire-and-forget POST
+                            let _ = reqwest::Client::new()
+                                .post(&reindex_url)
+                                .timeout(Duration::from_secs(5))
+                                .send()
+                                .await;
+                        }
+                    }
+                    KeyAction::None => {}
                 }
             }
         }
@@ -200,7 +378,28 @@ async fn run_remote_tui_loop(
 }
 
 // ---------------------------------------------------------------------------
-// Restore helpers
+// HTTP response types
+// ---------------------------------------------------------------------------
+
+#[derive(Debug, Deserialize)]
+struct InfoResponse {
+    chunks: usize,
+    files: usize,
+    max_chunk_id: u32,
+    db_size_human: String,
+    model: String,
+    dims: usize,
+    lock: String,
+    index_age: String,
+}
+
+#[derive(Debug, Deserialize)]
+struct DoctorResponse {
+    results: Vec<String>,
+}
+
+// ---------------------------------------------------------------------------
+// Terminal helpers
 // ---------------------------------------------------------------------------
 
 fn restore_terminal(terminal: &mut Terminal<CrosstermBackend<io::Stdout>>) -> io::Result<()> {
@@ -208,403 +407,4 @@ fn restore_terminal(terminal: &mut Terminal<CrosstermBackend<io::Stdout>>) -> io
     crossterm::execute!(io::stdout(), LeaveAlternateScreen)?;
     terminal.show_cursor()?;
     Ok(())
-}
-
-// ---------------------------------------------------------------------------
-// Key handling
-// ---------------------------------------------------------------------------
-
-fn is_quit_key(key: KeyEvent) -> bool {
-    matches!(key.code, KeyCode::Char('q'))
-}
-
-fn handle_key(key: KeyEvent, table_state: &mut TableState, row_count: usize) -> bool {
-    if row_count == 0 {
-        return false;
-    }
-    match key.code {
-        KeyCode::Up | KeyCode::Char('k') => {
-            let i = table_state.selected().unwrap_or(0);
-            if i > 0 {
-                table_state.select(Some(i - 1));
-            }
-        }
-        KeyCode::Down | KeyCode::Char('j') => {
-            let i = table_state.selected().unwrap_or(0);
-            if i < row_count - 1 {
-                table_state.select(Some(i + 1));
-            }
-        }
-        KeyCode::Home => {
-            table_state.select(Some(0));
-        }
-        KeyCode::End => {
-            table_state.select(Some(row_count - 1));
-        }
-        KeyCode::Char('s') => {
-            return true; // signal: reload requested
-        }
-        _ => {}
-    }
-    false
-}
-
-// ---------------------------------------------------------------------------
-// Rendering — mirrors the embedded TUI in tui.rs
-// ---------------------------------------------------------------------------
-
-fn render_header(f: &mut ratatui::Frame, area: Rect, serve_url: &str, version: &str) {
-    let now = chrono::Local::now().format("%H:%M:%S").to_string();
-
-    let title_line = Line::from(vec![
-        Span::styled(
-            format!(" codesearch serve v{} · ", version),
-            Style::default()
-                .fg(Color::Cyan)
-                .add_modifier(Modifier::BOLD),
-        ),
-        Span::styled(serve_url.to_string(), Style::default().fg(Color::White)),
-        Span::styled(format!("  {} ", now), Style::default().fg(Color::DarkGray)),
-        Span::styled("[remote]".to_string(), Style::default().fg(Color::Magenta)),
-    ]);
-
-    let block = Block::default()
-        .borders(Borders::TOP)
-        .border_style(Style::default().fg(Color::DarkGray));
-    let inner = block.inner(area);
-    f.render_widget(block, area);
-
-    let centered = Layout::vertical([Constraint::Length(1), Constraint::Min(0)]).split(inner);
-    f.render_widget(ratatui::widgets::Paragraph::new(title_line), centered[0]);
-}
-
-/// Returns true during the "bright" phase of a ~1s pulse cycle (500ms bright, 500ms dim).
-fn pulse_bright() -> bool {
-    std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_millis()
-        % 1000
-        < 500
-}
-
-fn render_table(
-    f: &mut ratatui::Frame,
-    area: Rect,
-    repos: &[RepoInfo],
-    table_state: &mut TableState,
-) {
-    let header_cells = [
-        "Alias",
-        "Status",
-        "Changes",
-        "Calls",
-        "Last Tool Call",
-        "Lock",
-    ];
-    let header = Row::new(
-        header_cells
-            .iter()
-            .map(|h| Cell::from(*h).style(Style::default().add_modifier(Modifier::BOLD))),
-    )
-    .style(Style::default().fg(Color::White))
-    .bottom_margin(1);
-
-    let max_alias_w = repos
-        .iter()
-        .map(|r| {
-            let extra = match r.csharp_index.as_str() {
-                "ready" => 4,    // " C#·"
-                "error" => 4,    // " C#!"
-                "indexing" => 4, // " C#…"
-                _ => 0,
-            };
-            r.alias.len() + extra
-        })
-        .max()
-        .unwrap_or(10)
-        .max(10);
-
-    let rows: Vec<Row> = repos
-        .iter()
-        .map(|repo| {
-            let status_cell = status_cell(&repo.status, &repo.csharp_index);
-            // Keep left edge stable: fixed-width, right-aligned, capped at 5 chars.
-            let changes_str = if repo.changes > 99999 {
-                " 99k+".to_string()
-            } else {
-                format!("{:>5}", repo.changes)
-            };
-            let changes_cell = Cell::from(changes_str).style(Style::default().fg(Color::White));
-            let calls_cell = if repo.tool_call_count > 0 {
-                Cell::from(format!("{:>5}", repo.tool_call_count))
-                    .style(Style::default().fg(Color::Cyan))
-            } else {
-                Cell::from("    -".to_string()).style(Style::default().fg(Color::DarkGray))
-            };
-            let tool_cell = Cell::from(repo.last_tool_call.as_deref().unwrap_or("—").to_string())
-                .style(Style::default().fg(Color::DarkGray));
-            let lock_cell = lock_cell(&repo.lock_mode);
-
-            // Alias cell with optional C# indicator
-            let alias_cell = match repo.csharp_index.as_str() {
-                "ready" => Cell::from(format!("{} C#·", repo.alias))
-                    .style(Style::default().fg(Color::White)),
-                "error" => {
-                    Cell::from(format!("{} C#!", repo.alias)).style(Style::default().fg(Color::Red))
-                }
-                "indexing" => {
-                    // Pulsing C# indicator during indexing (color only, fixed text width)
-                    if pulse_bright() {
-                        Cell::from(format!("{} C#…", repo.alias)).style(
-                            Style::default()
-                                .fg(Color::Yellow)
-                                .add_modifier(Modifier::BOLD),
-                        )
-                    } else {
-                        Cell::from(format!("{} C#…", repo.alias))
-                            .style(Style::default().fg(Color::DarkGray))
-                    }
-                }
-                _ => Cell::from(repo.alias.clone()).style(Style::default().fg(Color::White)),
-            };
-
-            // Red alias if the repo has errors
-            let alias_cell = if repo.status == "error" {
-                alias_cell.style(Style::default().fg(Color::Red).add_modifier(Modifier::BOLD))
-            } else {
-                alias_cell
-            };
-
-            Row::new(vec![
-                alias_cell,
-                status_cell,
-                changes_cell,
-                calls_cell,
-                tool_cell,
-                lock_cell,
-            ])
-        })
-        .collect();
-
-    let table = Table::new(
-        rows,
-        [
-            Constraint::Length(max_alias_w as u16 + 2),
-            Constraint::Length(12),
-            Constraint::Length(7),
-            Constraint::Length(7),
-            Constraint::Min(24),
-            Constraint::Length(7),
-        ],
-    )
-    .header(header)
-    .block(
-        Block::default()
-            .borders(Borders::NONE)
-            .style(Style::default().bg(Color::Reset)),
-    )
-    .row_highlight_style(
-        Style::default()
-            .bg(Color::Rgb(30, 30, 50))
-            .add_modifier(Modifier::BOLD),
-    )
-    .highlight_symbol("▶ ");
-
-    f.render_stateful_widget(table, area, table_state);
-}
-
-fn render_detail(f: &mut ratatui::Frame, area: Rect, repos: &[RepoInfo], table_state: &TableState) {
-    if repos.is_empty() {
-        return;
-    }
-
-    let idx = table_state.selected().unwrap_or(0);
-    if idx >= repos.len() {
-        return;
-    }
-
-    let repo = &repos[idx];
-
-    let status_label = match repo.status.as_str() {
-        "open" => "Open",
-        "warm" => "Warm (no FSW)",
-        "readonly" => "Readonly",
-        "closed" => "Closed",
-        "indexing" => "Indexing…",
-        "error" => "Error",
-        "no_index" => "No Index",
-        _ => &repo.status,
-    };
-
-    let detail_line = Line::from(vec![
-        Span::styled(" ▶ ", Style::default().fg(Color::Yellow)),
-        Span::styled(
-            repo.alias.clone(),
-            Style::default()
-                .fg(Color::White)
-                .add_modifier(Modifier::BOLD),
-        ),
-        Span::styled("  ", Style::default()),
-        Span::styled(status_label, Style::default().fg(Color::Cyan)),
-        Span::styled("  ", Style::default()),
-        Span::styled(
-            format!("lock: {}", repo.lock_mode),
-            Style::default().fg(Color::DarkGray),
-        ),
-    ]);
-
-    let tool_str = repo.last_tool_call.as_deref().unwrap_or("—");
-    let csharp_str = match repo.csharp_index.as_str() {
-        "ready" => "  C#·",
-        "error" => "  C#!",
-        "indexing" => "  C#…",
-        _ => "",
-    };
-    let csharp_color = match repo.csharp_index.as_str() {
-        "ready" => Color::Green,
-        "error" => Color::Red,
-        "indexing" => {
-            if pulse_bright() {
-                Color::Yellow
-            } else {
-                Color::DarkGray
-            }
-        }
-        _ => Color::DarkGray,
-    };
-    let info_line = Line::from(vec![
-        Span::styled("   changes:", Style::default().fg(Color::DarkGray)),
-        Span::styled(
-            format!(" {}  ", repo.changes),
-            Style::default().fg(Color::White),
-        ),
-        Span::styled("calls:", Style::default().fg(Color::DarkGray)),
-        Span::styled(
-            format!(" {}  ", repo.tool_call_count),
-            Style::default().fg(Color::Cyan),
-        ),
-        Span::styled("last:", Style::default().fg(Color::DarkGray)),
-        Span::styled(format!(" {}", tool_str), Style::default().fg(Color::White)),
-        Span::styled(csharp_str, Style::default().fg(csharp_color)),
-    ]);
-
-    let block = Block::default()
-        .borders(Borders::TOP)
-        .border_style(Style::default().fg(Color::Rgb(40, 40, 60)));
-    let inner = block.inner(area);
-    f.render_widget(block, area);
-
-    let detail_chunks =
-        Layout::vertical([Constraint::Length(1), Constraint::Length(1)]).split(inner);
-
-    f.render_widget(
-        ratatui::widgets::Paragraph::new(detail_line),
-        detail_chunks[0],
-    );
-    f.render_widget(
-        ratatui::widgets::Paragraph::new(info_line),
-        detail_chunks[1],
-    );
-}
-
-fn render_footer(
-    f: &mut ratatui::Frame,
-    area: Rect,
-    repos: &[RepoInfo],
-    table_state: &TableState,
-    active: u64,
-    cpu: &str,
-    csharp_helper: bool,
-) {
-    let selected = table_state.selected().unwrap_or(0);
-    let scroll_indicator = if repos.len() > 1 {
-        format!("[{}/{}]", selected + 1, repos.len())
-    } else {
-        String::new()
-    };
-
-    let sessions_str = format!("Sessions: {}", active);
-    let cpu_str = format!("CPU: {}", cpu);
-
-    let right_len = cpu_str.len() + sessions_str.len() + 3 + "C# │ ".len();
-
-    let footer_inner = area.inner(Margin {
-        vertical: 0,
-        horizontal: 1,
-    });
-    let [left, right] =
-        Layout::horizontal([Constraint::Min(0), Constraint::Length(right_len as u16 + 2)])
-            .areas(footer_inner);
-
-    let left_line = Line::from(vec![
-        Span::styled("[q] quit  ", Style::default().fg(Color::DarkGray)),
-        Span::styled("[↑↓] scroll  ", Style::default().fg(Color::DarkGray)),
-        Span::styled("[s] reload  ", Style::default().fg(Color::DarkGray)),
-        Span::styled(scroll_indicator, Style::default().fg(Color::Yellow)),
-    ]);
-
-    let csharp_indicator = if csharp_helper {
-        Span::styled("C# │ ", Style::default().fg(Color::Green))
-    } else {
-        Span::styled("C# │ ", Style::default().fg(Color::DarkGray))
-    };
-
-    let right_line = Line::from(vec![
-        csharp_indicator,
-        Span::styled(cpu_str, Style::default().fg(Color::Green)),
-        Span::styled(" │ ", Style::default().fg(Color::DarkGray)),
-        Span::styled(sessions_str, Style::default().fg(Color::Cyan)),
-    ]);
-
-    f.render_widget(ratatui::widgets::Paragraph::new(left_line), left);
-    f.render_widget(
-        ratatui::widgets::Paragraph::new(right_line).right_aligned(),
-        right,
-    );
-}
-
-// ---------------------------------------------------------------------------
-// Cell styling helpers
-// ---------------------------------------------------------------------------
-
-fn status_cell(status: &str, csharp: &str) -> Cell<'static> {
-    match status {
-        "open" => Cell::from("✓ ready   ".to_string()).style(
-            Style::default()
-                .fg(Color::Green)
-                .add_modifier(Modifier::BOLD),
-        ),
-        "warm" => match csharp {
-            "ready" => {
-                Cell::from("◐ warm+C# ".to_string()).style(Style::default().fg(Color::Green))
-            }
-            "indexing" => Cell::from("◐ warm C#…".to_string()).style(
-                Style::default()
-                    .fg(Color::Yellow)
-                    .add_modifier(Modifier::BOLD),
-            ),
-            "error" => Cell::from("◐ warm C#!".to_string()).style(Style::default().fg(Color::Red)),
-            _ => Cell::from("◐ warm    ".to_string()).style(Style::default().fg(Color::Yellow)),
-        },
-        "readonly" => Cell::from("◑ ro      ".to_string()).style(Style::default().fg(Color::Cyan)),
-        "indexing" => Cell::from("⟳ idx…    ".to_string()).style(
-            Style::default()
-                .fg(Color::Yellow)
-                .add_modifier(Modifier::BOLD),
-        ),
-        "closed" => Cell::from("○ closed  ".to_string()).style(Style::default().fg(Color::Gray)),
-        "error" => Cell::from("✗ error   ".to_string())
-            .style(Style::default().fg(Color::Red).add_modifier(Modifier::BOLD)),
-        "no_index" => Cell::from("— no idx  ".to_string()).style(Style::default().fg(Color::Gray)),
-        _ => Cell::from(format!("{:<10}", status)).style(Style::default().fg(Color::White)),
-    }
-}
-
-fn lock_cell(lock_mode: &str) -> Cell<'static> {
-    match lock_mode {
-        "write" => Cell::from("write".to_string()).style(Style::default().fg(Color::Cyan)),
-        "read" => Cell::from("read".to_string()).style(Style::default().fg(Color::White)),
-        _ => Cell::from("—".to_string()).style(Style::default().fg(Color::Gray)),
-    }
 }
